+++ Cargo.toml
@@ -14,3 +14,4 @@ serde_json = "*"
 log = "*"
 lazy_static = "1.3.0"
 lsp-types = { version = "0.57.1", features = ["proposed"] }
+simple-logging = "*"
