+++ src/lib.rs
@@ -3,3 +3,5 @@ pub type Result<T> = std::result::Result<T, Box<dyn std::error::Error + Send + S
 pub mod lspc;
 pub mod neovim;
 pub mod rpc;
+
+pub use lspc::Lspc;
